Move query workflow to api package
